% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-Docs.R
\docType{data}
\name{GreenBay_FoxRiver}
\alias{GreenBay_FoxRiver}
\title{WQP Green Bay Fox River Data from HUC12 040302040405}
\format{
A data frame of 14200 observations and 149 variables
}
\usage{
data(GreenBay_FoxRiver)
}
\description{
WQP Green Bay Fox River Data from HUC12 040302040405
}
