% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-Docs.R
\docType{data}
\name{WQP_with_ATTAINSonly}
\alias{WQP_with_ATTAINSonly}
\title{TADA_GetATTAINS Data From a Subset of WQP Green Bay Fox River Data from HUC12 040302040405}
\format{
A list containing: 
 TADA_with_ATTAINS, a tibble with 2877 rows and 188 columns
 ATTAINS_catchments, a data frame with 17 rows and 37 columns
 ATTAINS_points, NULL
 ATTAINS_lines, a data frame with 6 rows and 73 columns
 ATTAINS_polygons, a data frame with 5 rows and 74 columns
}
\usage{
data(WQP_with_ATTAINSonly)
}
\description{
TADA_GetATTAINS Data From a Subset of WQP Green Bay Fox River Data from HUC12 040302040405
}
